Vistas_ClientesCompras

+++ Examen/ExamenJoins.sql
@@ -1,5 +1,5 @@
 CREATE TABLE Autores (
-id_autor bigint identity(300,1) PRIMARY KEY,
+id_autor INT identity(300,1) PRIMARY KEY,
 nombre VARCHAR(100),
 pais_origen VARCHAR(100)
 );
@@ -7,7 +7,7 @@ pais_origen VARCHAR(100)
 select * from Autores
 
 ALTER TABLE Comics
-ADD id_autor bigint,
+ADD id_autor int,
 FOREIGN KEY (id_autor) REFERENCES Autores(id_autor);
 select * from Comics
